Factor out ConfigOperation class from Dispatcher
