+++ inbm/dispatcher-agent/tests/unit/test_dispatcher.py
@@ -3,13 +3,12 @@ import os
 from typing import Any
 from unittest import TestCase
 
-from mock import patch, Mock
+from unittest.mock import Mock, patch
 from dispatcher.dispatcher import WindowsDispatcherService
 from unit.common.mock_resources import *
 
 from dispatcher.aota.aota_error import AotaError
-from dispatcher.common.result_constants import PUBLISH_SUCCESS, CONFIG_LOAD_SUCCESS, CONFIG_LOAD_FAIL_WRONG_PATH, \
-    CODE_OK, CODE_BAD_REQUEST
+from dispatcher.common.result_constants import PUBLISH_SUCCESS, CODE_OK, CODE_BAD_REQUEST
 from dispatcher.constants import TargetType
 from dispatcher.dispatcher_class import Dispatcher
 from dispatcher.dispatcher_exception import DispatcherException
@@ -191,8 +190,7 @@ class TestDispatcher(TestCase):
         mock_ota_update.assert_called()
 
     @patch('inbm_lib.xmlhandler.XmlHandler', autospec=True)
-    @patch('inbm_lib.xmlhandler.XmlHandler.get_children')
-    def test_do_install_pota_do_ota_func_called(self, mock_get_children, MockXmlHandler, mock_logging) -> None:
+    def test_do_install_pota_do_ota_func_called(self, MockXmlHandler, mock_logging) -> None:
         parsed_head = MockXmlHandler.return_value
         resource = {'fota': ' ', 'sota': ' '}
         d = TestDispatcher._build_dispatcher()
@@ -205,7 +203,7 @@ class TestDispatcher(TestCase):
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.connect')
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.subscribe')
     @patch('dispatcher.dispatcher_class.Dispatcher._send_result')
-    @patch('dispatcher.dispatcher_class.Dispatcher._request_config_agent')
+    @patch('dispatcher.config.config_operation.ConfigOperation._request_config_agent')
     @patch('dispatcher.dispatcher_class.Dispatcher.invoke_workload_orchestration_check')
     def test_config_set_check_pass(self,
                                    mock_workload_orchestration_func: Any,
@@ -224,7 +222,7 @@ class TestDispatcher(TestCase):
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.connect')
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.subscribe')
     @patch('dispatcher.dispatcher_class.Dispatcher._send_result')
-    @patch('dispatcher.dispatcher_class.Dispatcher._request_config_agent')
+    @patch('dispatcher.config.config_operation.ConfigOperation._request_config_agent')
     @patch('dispatcher.dispatcher_class.Dispatcher.invoke_workload_orchestration_check')
     def test_config_set_check_fail(self,
                                    mock_workload_orchestration_func: Any,
@@ -361,7 +359,7 @@ class TestDispatcher(TestCase):
         mock_config_func.return_value = PUBLISH_SUCCESS
         self.assertEqual(200, d.do_install(xml=xml, schema_location=TEST_SCHEMA_LOCATION).status)
 
-    @patch('dispatcher.dispatcher_class.Dispatcher._do_config_install_load')
+    @patch('dispatcher.config.config_operation.ConfigOperation._do_config_install_load')
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.connect')
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.subscribe')
     @patch('dispatcher.dispatcher_class.Dispatcher._send_result')
@@ -384,7 +382,7 @@ class TestDispatcher(TestCase):
         mock_install_func.return_value = PUBLISH_SUCCESS
         self.assertEqual(200, d.do_install(xml=xml, schema_location=TEST_SCHEMA_LOCATION).status)
 
-    @patch('dispatcher.dispatcher_class.Dispatcher._do_config_install_load')
+    @patch('dispatcher.config.config_operation.ConfigOperation._do_config_install_load')
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.connect')
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.subscribe')
     @patch('dispatcher.dispatcher_class.Dispatcher._send_result')
@@ -404,47 +402,8 @@ class TestDispatcher(TestCase):
         self.assertIn("Error parsing/validating manifest: XML va", d.do_install(
             xml=xml, schema_location=TEST_SCHEMA_LOCATION).message)
 
-    @patch('dispatcher.dispatcher_class.Dispatcher._request_config_agent')
-    @patch('dispatcher.configuration_helper.ConfigurationHelper.parse_url')
-    @patch('dispatcher.configuration_helper.ConfigurationHelper.download_config')
-    def test_config_load_operation_on_local_path_pass(self,
-                                                      mock_download: Any,
-                                                      mock_url: Any,
-                                                      mock_req_conf_func: Any,
-                                                      mock_logging: Any) -> None:
 
-        xml = '<?xml version="1.0" encoding="UTF-8"?><manifest><type>config</type><config> ' \
-              '<cmd>load</cmd><configtype><load><path>/var/cache/manageability/intel.conf</path></load></configtype> ' \
-              '</config></manifest> '
-        parsed_head = XmlHandler(xml, is_file=False, schema_location=TEST_SCHEMA_LOCATION)
-        d = TestDispatcher._build_dispatcher()
-        # mock_xml_init.return_value = None
-        mock_url.return_value = None
-        mock_download.return_value = False, None
-        mock_req_conf_func.assert_not_called()
-        self.assertEqual(CONFIG_LOAD_SUCCESS, d._do_config_install_load(
-            parsed_head=parsed_head, xml=xml))
 
-    @patch('dispatcher.dispatcher_class.Dispatcher._request_config_agent')
-    @patch('dispatcher.configuration_helper.ConfigurationHelper.parse_url')
-    @patch('dispatcher.configuration_helper.ConfigurationHelper.download_config')
-    def test_config_load_operation_on_local_path_fail(self,
-                                                      mock_download: Any,
-                                                      mock_url: Any,
-                                                      mock_req_conf_func: Any,
-                                                      mock_logging: Any) -> None:
-
-        xml = '<?xml version="1.0" encoding="UTF-8"?><manifest><type>config</type><config> ' \
-              '<cmd>load</cmd><configtype><load><path>/var/cache/abc/intel.conf</path></load></configtype> ' \
-              '</config></manifest> '
-        parsed_head = XmlHandler(xml, is_file=False, schema_location=TEST_SCHEMA_LOCATION)
-        d = TestDispatcher._build_dispatcher()
-        # mock_xml_init.return_value = None
-        mock_url.return_value = None
-        mock_download.return_value = False, None
-        mock_req_conf_func.assert_not_called()
-        self.assertEqual(CONFIG_LOAD_FAIL_WRONG_PATH, d._do_config_install_load(
-            parsed_head=parsed_head, xml=xml))
 
     @patch('dispatcher.dispatcher_class.Dispatcher.invoke_workload_orchestration_check')
     @patch('dispatcher.dispatcher_class.Dispatcher._perform_cmd_type_operation')
@@ -473,7 +432,7 @@ class TestDispatcher(TestCase):
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.connect')
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.subscribe')
     @patch('dispatcher.dispatcher_class.Dispatcher._send_result')
-    @patch('dispatcher.dispatcher_class.Dispatcher._request_config_agent')
+    @patch('dispatcher.config.config_operation.ConfigOperation._request_config_agent')
     @patch('dispatcher.dispatcher_class.Dispatcher.invoke_workload_orchestration_check')
     def test_config_get_element_fail(self,
                                      mock_workload_orchestration: Any,
@@ -491,7 +450,7 @@ class TestDispatcher(TestCase):
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.connect')
     @patch('inbm_lib.mqttclient.mqtt.mqtt.Client.subscribe')
     @patch('dispatcher.dispatcher_class.Dispatcher._send_result')
-    @patch('dispatcher.dispatcher_class.Dispatcher._request_config_agent')
+    @patch('dispatcher.config.config_operation.ConfigOperation._request_config_agent')
     @patch('dispatcher.dispatcher_class.Dispatcher.invoke_workload_orchestration_check')
     def test_config_get_element_pass(self,
                                      mock_workload_orchestration: Any,
